Aplikacja_Webowa_Etap_4/serwer/requirements.txt
flask
redis
pyjwt